+++ app/src/main/res/layout/Com1314080901142Activity.xml
@@ -1,0 +1,18 @@
+   <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
+        android:layout_height="match_parent" android:paddingLeft="@dimen/activity_horizontal_margin"
+        android:paddingRight="@dimen/activity_horizontal_margin"
+        android:paddingTop="@dimen/activity_vertical_margin"
+        android:paddingBottom="@dimen/activity_vertical_margin" tools:context=".MainActivity">
+
+        <TextView android:text="@string/hello_world" android:layout_width="wrap_content"
+           android:layout_height="wrap_content" />
+        <ImageView
+           android:layout_width="wrap_content"
+           android:layout_height="" />
+        <ImageView
+            android:layout_width="50dp"
+            android:layout_height="50dp"
+          android:src="@drawable/laven"/>
+
+    </RelativeLayout>
